First commit and last.
This is my first and last commit for this module. I wrote this for a specific use case and probably wont ever come back around to making it...better. Feel free to change it all. This might not even work as it is because I only used it a few times. It works fine with multiple servers added to the JSON file. If you need to restore to specific directories, you'll need to update the code to make that more modular.